Modifications to reach 95 in rubycritic
