+++ bin/grid.rb
@@ -1,11 +1,10 @@
 # frozen_string_literal: true
 
 require 'colorize'
-require_relative 'maze_level_two'
-require_relative 'maze_level_one'
+require_relative 'maze_levels'
 
 # This is the class to create a grid to the game
-class Grid
+class Grid < MazeLevels
   GAME_ELEMENTS = {
     horizontal: '▀▀▀'.blue,
     vertical: ' █ '.blue,
@@ -21,16 +20,6 @@ class Grid
     @width = width
     @height = height
     @grid = Array.new(height) { Array.new(width) { :food } }
-  end
-
-  def generate_maze
-    laberinto = MazeLevelOne.new(@grid)
-    laberinto.create_maze(@width - 1, @height - 1)
-  end
-
-  def generate_maze_level_two
-    laberinto = MazeLevelTwo.new(@grid)
-    laberinto.create_maze(@width - 1, @height - 1)
   end
 
   def show_maze(pacman)
